introduce seperate replace-and-select command
