+++ addon/commands/index.ts
@@ -1,0 +1,1 @@
+export { default as replaceSelectionWithAndSelectNode } from './replace-selection-and-select-node';
