+++ addon/plugins/variable-plugin/variables/codelist.ts
@@ -166,6 +166,7 @@ const emberNodeConfig: EmberNodeConfig = {
   uriAttributes: ['variableInstance'],
   draggable: false,
   needsFFKludge: true,
+  selectable: true,
   attrs: {
     ...rdfaAttrSpec,
     selectionStyle: {
